pyproject.toml
[tool.poetry]
name = "nonebot-plugin-csgo-case-simulator"
version = "0.1.8"
description = "a nonebot based csgo case simulator"
authors = ["Roy <lyt2980999208@gmail.com>"]
license = "MIT"
packages = [{ include = "nonebot-plugin-csgo-case-simulator" }]

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = "^2.0.0-rc.4"
nonebot-adapter-onebot = "^2.2.3"
Pillow = "^9.5.0"
httpx = "^0.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
